Updated marquee & icons
Scrolling text is now pulled from a text file titled "mynews.txt" and
stored in an array titled "marqueeData". A random element of the array
is stored in the variable cSentence, which is diplayed in the marquee.
cSentence is updated every 30 seconds with a different random element
from the array.

Additionally, the icons representing inventory have been updated.